requirements.txt
discord.py
python-dotenv
gspread
numpy
pandas
requests
supabase